+++ PcSupport/ConfigPc.cpp
@@ -24,7 +24,7 @@ bool ConfigPc::SaveUserConfig()
 		return false;
 	}
 
-	for( int i = 0; i < rConfig.GetConfigObjectCount(); ++i )
+	for( size_t i = 0; i < rConfig.GetConfigObjectCount(); ++i )
 	{
 		const Name &name = rConfig.GetConfigObjectName( i );
 		FilePath path = rConfig.GetUserConfigObjectFilePath( name );
